gradio-adk-agent/pyproject.toml
[project]
name = "keyword-agent-gradio-adk"
version = "0.1.0"
description = "Keyword- & SEO-Agent (Google ADK) mit Gradio UI und optionaler Trends-Anreicherung"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.20.0",
    "google-genai>=0.3.0",
    "gradio>=6.1.0",
    "pandas>=2.0.0",
    "altair>=5.0.0",
    "python-dotenv>=1.0.0",
    "pytrends>=4.9.0",
]